tests/wpt/web-platform-tests/html/semantics/edits/the-del-element/del_effect.html

<!DOCTYPE html>
<meta charset=UTF-8>
<title>HTML Test: Text in the del element should be 'line-through'</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#the-del-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p><del>crossed-off text</del></p>
<div id="log"></div>

<script>
  test(function() {
    var element = document.getElementsByTagName('del')[0],
        textDecoration = getComputedStyle(element).textDecorationLine ||
                         getComputedStyle(element).textDecoration;
    assert_equals(textDecoration, 'line-through');
  });
</script>
